REFACTOR:: rename module folder
