+++ gsheets/src/authentication.py
@@ -2,7 +2,7 @@ from pathlib import Path
 from google.oauth2.credentials import Credentials
 from google.auth.transport.requests import Request
 from google_auth_oauthlib.flow import InstalledAppFlow
-from authentication.utils import run_with_timeout
+from src.utils import run_with_timeout
 import logging
 
 CREDENTIALS_FILE_NAME = "credentials.json"
