+++ gsheets/src/tests/test_authentication.py
@@ -1,14 +1,14 @@
 import pytest
 from unittest.mock import patch, MagicMock, mock_open
 from pathlib import Path
-from authentication.authentication import (
+from src.authentication import (
     Authentication,
     folder_contains_credentials_file,
     folder_contains_token_file,
 )
 
 
-@patch("authentication.authentication.folder_contains_credentials_file")
+@patch("src.authentication.folder_contains_credentials_file")
 def test_init_with_not_existing_credentials(mocked_folder_contains_credentials_file):
     # Given
     mocked_folder_contains_credentials_file.return_value = False
@@ -23,7 +23,7 @@ def test_init_with_not_existing_credentials(mocked_folder_contains_credentials_f
     )
 
 
-@patch("authentication.authentication.folder_contains_credentials_file")
+@patch("src.authentication.folder_contains_credentials_file")
 def test_init_with_existing_credentials(mocked_folder_contains_credentials_file):
     # Given
     mocked_folder_contains_credentials_file.return_value = True
@@ -42,7 +42,7 @@ def test_init_with_existing_credentials(mocked_folder_contains_credentials_file)
     assert str(test_auth.path_to_token) == "/path/to/existing/credentials/token.json"
 
 
-@patch("authentication.authentication.Path")
+@patch("src.authentication.Path")
 def test_folder_contains_credentials_file(mocked_path):
     # When
     folder_contains_credentials_file("/path/to/credentials.json")
@@ -51,7 +51,7 @@ def test_folder_contains_credentials_file(mocked_path):
     mocked_path.assert_called_once_with("/path/to/credentials.json")
 
 
-@patch("authentication.authentication.Path")
+@patch("src.authentication.Path")
 def test_folder_contains_token_file(mocked_path):
     # When
     folder_contains_token_file("/path/to/token.json")
@@ -63,7 +63,7 @@ def test_folder_contains_token_file(mocked_path):
 @pytest.fixture
 def auth_fixture():
     with patch(
-        "authentication.authentication.folder_contains_credentials_file"
+        "src.authentication.folder_contains_credentials_file"
     ) as mocked_folder_contains_credentials_file:
         mocked_folder_contains_credentials_file.return_value = True
         test_auth = Authentication(credentials_folder="/path/to/existing/credentials")
@@ -101,7 +101,7 @@ def test_check_with_different_tokens(
     test_auth = auth_fixture[0]
     token_path = auth_fixture[1].joinpath("token.json")
 
-    with patch("authentication.authentication.Credentials") as MockCredentials:
+    with patch("src.authentication.Credentials") as MockCredentials:
         mock_credentials = MagicMock()
         mock_credentials.expired = expired
         mock_credentials.refresh_token = refresh_token
@@ -110,11 +110,11 @@ def test_check_with_different_tokens(
         MockCredentials.from_authorized_user_file.return_value = mock_credentials
 
         with patch(
-            "authentication.authentication.folder_contains_token_file"
+            "src.authentication.folder_contains_token_file"
         ) as mocked_folder_contains_token_file:
             mocked_folder_contains_token_file.return_value = True
 
-            with patch("authentication.authentication.logging") as mocked_logging:
+            with patch("src.authentication.logging") as mocked_logging:
                 # when
                 return_value = test_auth.check_credentials()
 
@@ -130,7 +130,7 @@ def test_check_with_not_existing_token(auth_fixture):
     token_path = auth_fixture[1].joinpath("token.json")
 
     with patch(
-        "authentication.authentication.folder_contains_token_file"
+        "src.authentication.folder_contains_token_file"
     ) as mocked_folder_contains_token_file:
         mocked_folder_contains_token_file.return_value = False
 
@@ -146,9 +146,7 @@ def test_successful_login(auth_fixture):
     # Given
     test_auth = auth_fixture[0]
 
-    with patch(
-        "authentication.authentication.InstalledAppFlow"
-    ) as MockedInstalledAppFlow:
+    with patch("src.authentication.InstalledAppFlow") as MockedInstalledAppFlow:
         mocked_flow = MagicMock()
         mocked_flow.run_local_server.return_value = MagicMock()
         MockedInstalledAppFlow.from_client_secrets_file.return_value = mocked_flow
@@ -168,16 +166,12 @@ def test_login_timeout(auth_fixture):
     # Given
     test_auth = auth_fixture[0]
 
-    with patch(
-        "authentication.authentication.InstalledAppFlow"
-    ) as MockedInstalledAppFlow:
+    with patch("src.authentication.InstalledAppFlow") as MockedInstalledAppFlow:
         mocked_flow = MagicMock()
         mocked_flow.run_local_server.return_value = None
         MockedInstalledAppFlow.from_client_secrets_file.return_value = mocked_flow
 
-        with patch(
-            "authentication.authentication.run_with_timeout"
-        ) as mocked_run_with_timeout:
+        with patch("src.authentication.run_with_timeout") as mocked_run_with_timeout:
             mocked_run_with_timeout.return_value = None
 
             # When
@@ -204,7 +198,7 @@ def test_saving_of_credentials(auth_fixture):
 
     test_auth.credentials = mock_credentials
 
-    with patch("authentication.authentication.open", mock_open()) as mocked_open:
+    with patch("src.authentication.open", mock_open()) as mocked_open:
         # When
         test_auth.save_credentials()
 
@@ -216,7 +210,7 @@ def test_saving_of_credentials(auth_fixture):
 def test_saving_of_empty_credentials(auth_fixture):
     # Given
     test_auth = auth_fixture[0]
-    with patch("authentication.authentication.open", mock_open()) as mocked_open:
+    with patch("src.authentication.open", mock_open()) as mocked_open:
         # When
         test_auth.save_credentials()
